requirements.txt
aiogram==3.*
python-dotenv
python-telegram-bot==20.7